python/requirements.txt
linkedin-jobs-scraper==0.1.0
boto3==1.34.69
opensearch-py==2.4.2
requests-aws4auth==1.2.3
python-dotenv==1.0.1 